Added the IPipeline interface to implement the decode and execution of the instructions.

+++ SharpBoy/core/Memory.cs
@@ -8,7 +8,7 @@ namespace SharpBoy.Core
 {
     public class Memory : IMemory
     {
-        byte IMemory.this[int address] { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        byte IMemory.this[short address] { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         int IMemory.Size { get => throw new NotImplementedException(); }
     }
 }
